requirements.txt
# Core Web Framework
fastapi==0.116.0
uvicorn[standard]==0.35.0

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.2

# LangChain Stack - Updated to v0.3.x (Breaking changes - see migration guide)
langsmith>=0.4.0,<0.5.0  # Updated for compatibility with new LangChain
langchain-core>=0.3.0   # Major version update
langchain>=0.3.26        # Major version update
langchain-openai==0.3.27
langchain-community>=0.3.27
langchain-huggingface==0.3.0  # New package for HuggingFace integration
langchain-ollama==0.3.4       # New package for Ollama integration
langgraph==0.5.1

# LLM Providers - Updated versions
transformers>=4.53.1  # Latest version
torch>=2.0.0

# Hugging Face Integration
huggingface-hub>=0.23.0  # Updated for compatibility
accelerate>=0.25.0  # For faster model loading/inference
# langchain-huggingface==0.0.1  # Commented out due to dependency conflicts - now using 0.3.0 above

# Optional: For running larger models locally
# bitsandbytes==0.41.3  # For 8-bit/4-bit quantization
# optimum==1.16.1      # For model optimization

# Vector Stores and Embeddings
chromadb==1.0.15  # Major version update - check for breaking changes
sentence-transformers==5.0.0  # Major version update - check for breaking changes
tiktoken==0.9.0
faiss-cpu  # For efficient vector similarity search
faiss-gpu  # Use instead of faiss-cpu for GPU acceleration (requires CUDA)

# Data Processing
pandas>=2.1.0
numpy>=1.26.4  # Stay on 1.x for compatibility - NumPy 2.x may have breaking changes
scikit-learn>=1.3.0
nltk==3.8.1
spacy==3.7.2

# Web Scraping (for LEGO data collection)
requests>=2.31.0
beautifulsoup4==4.13.4
scrapy==2.13.3

# Data Validation
pydantic==2.11.7
pydantic-settings==2.10.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20

# Environment Management
python-dotenv==1.1.1

# Logging and Monitoring
loguru==0.7.3

# Testing
pytest>=7.4.0
pytest-asyncio==1.0.0
httpx==0.28.1

# Development Tools
black==25.1.0
flake8==7.3.0
isort==6.0.1
pre-commit==4.2.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.6.15

# Jupyter (for data exploration)
jupyter>=1.0.0
ipykernel>=6.0.0

# Data Visualization (optional, for development/analysis)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly==6.2.0

# File Processing
openpyxl==3.1.5
PyYAML>=6.0.0

# Rate Limiting (for API)
slowapi==0.1.9

# Background Tasks
celery==5.5.3
redis==6.2.0

# Production WSGI Server (alternative to uvicorn)
gunicorn==23.0.0

# Ollama
ollama>=0.5.1  